version 1.2.2:
  - Change the auto running mechanism
  - Add some public methods in the class Yod_Request
  - Add a public method select in the class Yod_Model
  - Add a public method field in the class Yod_DbModel
  - Change the method select of class Yod_DbModel
  - Add an abstract method count in the class Yod_Database
  - Add a public method count in the class Yod_DbPdo
  - Change the method find of class Yod_Model an Yod_DbModel

+++ yod/yod_request.h
@@ -28,7 +28,7 @@
 
 extern zend_class_entry *yod_request_ce;
 
-zval *yod_request_construct(zval *object, char *route, size_t route_len TSRMLS_DC);
+zval *yod_request_construct(zval *object, char *route, uint route_len TSRMLS_DC);
 int yod_request_dispatch(yod_request_t *object TSRMLS_DC);
 void yod_request_error404(yod_request_t *object, zval *html TSRMLS_DC);
 void yod_request_erroraction(yod_request_t *object TSRMLS_DC);
